backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
babel==2.17.0
beautifulsoup4==4.14.2
billiard==4.2.1
black==25.1.0
bleach==6.2.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
comm==0.2.3
constantly==23.10.4
contourpy==1.3.3
coverage==7.9.2
crispy-bootstrap5==2025.6
cryptography==45.0.4
cycler==0.12.1
daphne==4.2.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
dj-database-url==3.0.0
dj-rest-auth==7.0.1
Django==5.1.3
django-allauth==65.9.0
django-axes==8.0.0
django-ckeditor==6.7.3
django-cleanup==9.0.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-js-asset==3.1.2
django-mptt==0.17.0
django-redis==6.0.0
django-rest-knox==5.0.2
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-yasg==1.21.10
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.19.1
flake8==7.3.0
flower==2.0.1
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.7.0
google-auth==2.40.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
hyperlink==21.0.0
identify==2.6.14
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
isort==6.0.1
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterthemes==0.20.0
kiwisolver==1.4.9
kombu==5.5.4
lark==1.3.0
lesscpy==0.15.1
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.1.4
msgpack==1.1.1
multidict==6.6.4
mypy==1.17.0
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nodeenv==1.9.1
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.2.5
oauthlib==3.3.1
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
pluggy==1.6.0
ply==3.11
pre_commit==4.3.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.5
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
redis==6.2.0
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
rsa==4.9.1
Send2Trash==1.8.3
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.2
stack-data==0.6.3
Telethon==1.40.0
terminado==0.18.1
tinycss2==1.4.0
tomli==2.2.1
tornado==6.5.1
traitlets==5.14.3
Twisted==25.5.0
txaio==25.6.1
types-python-dateutil==2.9.0.20250822
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.32.1
vine==5.1.0
virtualenv==20.34.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
whitenoise==6.8.2
yarl==1.20.1
zope.interface==7.2
